% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{aminoacids}
\alias{aminoacids}
\title{Amminoacids data set.}
\format{A `data.frame` with 11 columns (OneLetter, ThreeLetter, FullName,
AvgMass, MonoMass, H, C, N, O S, Se) for the aminoacids. The H/C/N/O/S/Se
columns contain the number of elements that build the aminoacid.}
\source{
Taken from the unimod database: http://www.unimod.org/xml/unimod.xml.
}
\usage{
aminoacids
}
\description{
`data.frame` of aminoacids from the unimod database.
}
\details{
It was created as
follows:

```
unimod:::.createDataSets()
```
}
\examples{
data(aminoacids)
head(aminoacids)
}
\keyword{datasets}
