% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{elements}
\alias{elements}
\title{Elements data set.}
\format{
A `data.frame` with 4 columns (Name, FullName, AvgMass, MonoMass) for
the chemical elements.
}
\source{
Taken from the unimod database: http://www.unimod.org/xml/unimod.xml.
}
\usage{
elements
}
\description{
`data.frame` of chemical elements from the unimod database.
}
\details{
It was created as
follows:

```
unimod:::.createDataSets()
```
}
\examples{
data(elements)
head(elements)
}
\keyword{datasets}
