% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Modification-class}
\alias{Modification-class}
\title{The Modification Class}
\value{
TODO
}
\description{
The Modification Class
}
\details{
TODO
}
\section{Slots}{

\describe{
\item{\code{TODO}}{}
}}

\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
