% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{unimod-package}
\alias{unimod-package}
\title{Managing amino acid modifications for mass spectrometry in R.}
\description{
An interface to the community supported database for amino
acid/protein modifications using mass spectrometry.
}
\details{
Unimod is a public domain database, distributed under a copyleft
licence: 'a copyright notice that permits unrestricted redistribution
and modification, provided that all copies and derivatives retain the
same permissions.'

The aim is to create a community supported, comprehensive database of
protein modifications for mass spectrometry applications. That is,
accurate and verifiable values, derived from elemental compositions,
for the mass differences introduced by all types of natural and
artificial modifications. Other important information includes any
mass change, (neutral loss), that occurs during MS/MS analysis, and
site specificity, (which residues are susceptible to modification and
any constraints on the position of the modification within the protein
or peptide).

The goal is not to place this information in a biological
context. There are already many high quality sources of biological
information about post-translational modifications:

\itemize{
 \item{RESID for post-translational modifications}
 \item{Swiss-Prot protein sequence database}
 \item{Prosite database of protein families and domains}
 \item{Glycan Database from the Consortium for Functional Glycomics}
}

Other compilations of modifications with a mass spectrometry bias:

\itemize{
 \item{Delta Mass}
 \item{FindMod}
}
}
\references{
Source and more details: \url{http://www.unimod.org/unimod_help.html}

\url{https://github.com/ComputationalProteomicsUnit/unimod/}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\keyword{package}
