% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/functions-Modification.R,
%   R/methods-Modification.R
\docType{class}
\name{Modification-class}
\alias{Modification-class}
\alias{Modification}
\alias{accession}
\alias{id}
\alias{accessions}
\alias{accessions,Modification-method}
\alias{composition}
\alias{deltaAvgMass}
\alias{deltaMonoMass}
\alias{name}
\alias{references}
\alias{specificity}
\alias{dim,Modification-method}
\alias{length,Modification-method}
\alias{names,Modification-method}
\title{Modification class}
\usage{
Modification(...)

accessions(object, ...)

composition(object)

deltaAvgMass(object)

deltaMonoMass(object)

name()

references(object)

specificity(object, hidden = TRUE, ...)

\S4method{dim}{Modification}(x)

\S4method{length}{Modification}(x)

\S4method{names}{Modification}(x)
}
\arguments{
\item{object}{Modification}

\item{hidden}{logical, show hidden specificities (default: TRUE)?}

\item{x}{Modification object}

\item{\ldots}{arguments passed to the internal constructor.}
}
\value{
integer, accession number/id.

double, named vector

double, delta average mass

double, delta monoisotopic mass

character, PSI name

data.frame, with

data.frame, with specificities
}
\description{
Modification class
}
\section{Functions}{
\itemize{
\item \code{Modification}: Constructor

\item \code{deltaAvgMass}: mass

\item \code{deltaMonoMass}: mass

\item \code{name}: name

\item \code{references}: references

\item \code{specificity}: specificity
}}

\section{Slots}{

\describe{
\item{\code{id}}{integer, unimod record_id.}

\item{\code{name}}{character, unimod PSI-MS name/title.}

\item{\code{description}}{character, unimod full_name.}

\item{\code{lastModified}}{character, time of last modification.}

\item{\code{approved}}{logical, approved?}

\item{\code{deltaAvgMass}}{double, unimod delta avge_mass.}

\item{\code{deltaMonoMass}}{double, unimod delta mono_mass.}

\item{\code{composition}}{integer, named vector. The names are the titles of unimod
elements. The items are the number of the elements.}

\item{\code{specificity}}{data.frame, with site, position and classification of the
modification.}

\item{\code{refs}}{data.frame, references listed in the unimod database.}
}}

