% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/functions-Modification.R
\docType{class}
\name{Modification-class}
\alias{Modification-class}
\alias{Modification}
\title{Modification class}
\usage{
Modification(...)
}
\description{
Modification class
}
\section{Functions}{
\itemize{
\item \code{Modification}: Constructor
}}

\section{Slots}{

\describe{
\item{\code{id}}{integer, unimod record_id.}

\item{\code{name}}{character, unimod PSI-MS name/title.}

\item{\code{description}}{character, unimod full_name.}

\item{\code{lastModified}}{character, time of last modification.}

\item{\code{approved}}{logical, approved?}

\item{\code{deltaAvgMass}}{double, unimod delta avge_mass.}

\item{\code{deltaMonoMass}}{double, unimod delta mono_mass.}

\item{\code{composition}}{integer, named vector. The names are the titles of unimod
elements. The items are the number of the elements.}

\item{\code{specificity}}{data.frame, with site, position and classification of the
modification.}

\item{\code{refs}}{data.frame, references listed in the unimod database.}
}}

