% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Modification-class}
\alias{Modification-class}
\title{Modification class}
\description{
Modification class
}
\section{Slots}{

\describe{
\item{\code{title}}{character, unimod title.}

\item{\code{name}}{character, unimod full_name.}

\item{\code{id}}{integer, unimod record_id.}

\item{\code{lastModified}}{character, time of last modification.}

\item{\code{deltaAvgMass}}{double, unimod delta avge_mass.}

\item{\code{deltaMonoMass}}{double, unimod delta mono_mass.}

\item{\code{composition}}{integer, named vector. The names are the titles of unimod
elements. The items are the number of the elements.}

\item{\code{specifity}}{data.frame, with site, position and classification of the
modification.}

\item{\code{refs}}{data.frame, references listed in the unimod database.}
}}

