% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{modifications}
\alias{modifications}
\title{Modifications data set.}
\format{
A `data.frame` with 15 columns (Id (created by
Name:(Position-)Site(:NeutralLoss) because unimod id is not unique),
UnimodId, Name, Description, AvgMass, MonoMass, Site, Position,
Classification, SpecGroup, NeutralLoss, LastModification, Approved,
Hidden) for the modifications.
}
\source{
Taken from the unimod database: http://www.unimod.org/xml/unimod.xml.
}
\usage{
modifications
}
\description{
`data.frame` of modifications from the unimod database.
}
\details{
It was created as
follows:

```
unimod:::.createDataSets()
```
}
\examples{
data(modifications)
head(modifications)
}
\keyword{datasets}
