% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{umodId}
\alias{umodId}
\alias{umodTitle}
\alias{umodName}
\alias{umodTitle}
\title{Find Modification by id/name/title.}
\usage{
umodId(id)

umodName(name)

umodTitle(title)
}
\arguments{
\item{id}{integer, id of the modification.}

\item{name}{character, PSI-MS name/title of the modification.}

\item{title}{character, PSI-MS name/title of the modification.}
}
\value{
object of class Modification.
}
\description{
This function allows to query the unimod database by the \emph{id} or
\emph{title} of specific modification and returns an
\code{\linkS4class{Modification}} object.
}
\examples{
library("unimod")

umodId(1)

umodTitle("Acetyl")

}
\references{
\url{http://www.unimod.org/modifications_list.php}
}
\seealso{
\code{\linkS4class{Modification}}.
}
